+++ barcode_detection/boundingbox/bounding_box.py
