+++ barcode_detection/decoding/decode.py
@@ -1,8 +1,10 @@
-from abc import ABC, abstractmethod
 import numpy as np
+
+from abc import ABC, abstractmethod
+from barcode_detection.boundingbox.bounding_box import BoundingBox
 
 
 class Decoder(ABC):
     @abstractmethod
-    def decode(self, input_img: np, bounding_boxes):
+    def decode(self, input_img: np.ndarray, bounding_boxes: list[BoundingBox]):
         pass
