+++ barcode_detection/localization/localize.py
@@ -1,8 +1,10 @@
-from abc import ABC, abstractmethod
 import numpy as np
+
+from abc import ABC, abstractmethod
+from barcode_detection.boundingbox.bounding_box import BoundingBox
 
 
 class Localizer(ABC):
     @abstractmethod
-    def get_boundings(self, input_img: np):
+    def get_boundings(self, input_img: np.ndarray) -> list[BoundingBox]:
         pass
