+++ barcode_detection/localization/localize_yolo.py
@@ -1,7 +1,7 @@
 import cv2
 import numpy as np
 
-from barcode_detection.boundingbox.BoundingBox import BoundingBox
+from barcode_detection.boundingbox.bounding_box import BoundingBox
 from barcode_detection.localization.localize import Localizer
 from barcode_detection.localization.onnx_yolov7 import OnnxDetector
